Faster rendering by Cython, moved *.py to src/*.py
