+++ examples/file_example_2.sivf
@@ -1,36 +1,261 @@
 /**
  * This is SIVF file :) 
- * Example 2
+ * Example 3
  */
 
 {
-    "sizes_wh": ["512", "512"],
+    "sizes_wh": ["1152", "1152"],
     "color_scheme": "rgb",
 
     "image": {
         "layer1": {
-            "blending": ["add", "overlap"], 
+            "square1": {
+                "xy": ["0", "0"],
+                "side": "1152",
+                "color": "#ff000000"
+            },
             "circle1": {
-                "xy": ["-25.981%", "-15%"],
-                "r": "43.589%",
-                "color": "#ffff0000"
+                "xy": ["0", "0"],
+                "r": "576",
+                "color": "#ffffffff"
+            }
+        },
+
+
+
+        "layer2": {
+            "square1": {
+                "xy": ["384", "0"],
+                "side": "384",
+                "color": "#ff202020"
+            },
+            "circle1": {
+                "xy": ["384", "0"],
+                "r": "192",
+                "color": "#ffe0e0e0"
+            },
+
+            "square2": {
+                "xy": ["0", "384"],
+                "side": "384",
+                "color": "#ff202020"
             },
             "circle2": {
-                "xy": ["0%", "30%"],
-                "r": "43.589%",
-                "color": "#ff00ff00"
+                "xy": ["0", "384"],
+                "r": "192",
+                "color": "#ffe0e0e0"
+            },
+
+            "square3": {
+                "xy": ["-384", "0"],
+                "side": "384",
+                "color": "#ff202020"
             },
             "circle3": {
-                "xy": ["25.981%", "-15%"],
-                "r": "43.589%",
-                "color": "#ff0000ff"
+                "xy": ["-384", "0"],
+                "r": "192",
+                "color": "#ffe0e0e0"
+            },
+            
+            "square4": {
+                "xy": ["0", "-384"],
+                "side": "384",
+                "color": "#ff202020"
             },
             "circle4": {
-                "inverse": "true",
-                "xy": ["0", "0"],
-                "r": "50%",
-                "color": "#00000000"
+                "xy": ["0", "-384"],
+                "r": "192",
+                "color": "#ffe0e0e0"
             }
+        },
+
+
+
+        "layer3": {
+            "square1": {
+                "xy": ["512", "0"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle1": {
+                "xy": ["512", "0"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square2": {
+                "xy": ["256", "0"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle2": {
+                "xy": ["256", "0"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square3": {
+                "xy": ["384", "128"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle3": {
+                "xy": ["384", "128"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square4": {
+                "xy": ["384", "-128"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle4": {
+                "xy": ["384", "-128"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+
+            
+            "square5": {
+                "xy": ["128", "384"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle5": {
+                "xy": ["128", "384"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square6": {
+                "xy": ["0", "512"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle6": {
+                "xy": ["0", "512"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square7": {
+                "xy": ["-128", "384"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle7": {
+                "xy": ["-128", "384"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square8": {
+                "xy": ["0", "256"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle8": {
+                "xy": ["0", "256"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+
+            
+            "square9": {
+                "xy": ["-256", "0"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle9": {
+                "xy": ["-256", "0"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square10": {
+                "xy": ["-384", "128"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle10": {
+                "xy": ["-384", "128"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square11": {
+                "xy": ["-512", "0"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle11": {
+                "xy": ["-512", "0"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square12": {
+                "xy": ["-384", "-128"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle12": {
+                "xy": ["-384", "-128"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+
+            
+            "square13": {
+                "xy": ["128", "-384"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle13": {
+                "xy": ["128", "-384"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square14": {
+                "xy": ["0", "-256"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle14": {
+                "xy": ["0", "-256"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square15": {
+                "xy": ["-128", "-384"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle15": {
+                "xy": ["-128", "-384"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            },
+
+            "square16": {
+                "xy": ["0", "-512"],
+                "side": "128",
+                "color": "#ff404040"
+            },
+            "circle16": {
+                "xy": ["0", "-512"],
+                "r": "64",
+                "color": "#ffc0c0c0"
+            }
+
+
+            
         }
     }
 }
